requirements.txt
customtkinter
tkcalendar
psycopg2
reportlab
PyMuPDF
Pillow
pywin32
pyinstaller